Introduced JDT Facade project
