+++ org.eclipse.xtext.java/META-INF/MANIFEST.MF
@@ -10,7 +10,8 @@ Require-Bundle: org.eclipse.xtext;bundle-version="2.37.0",
  org.eclipse.xtend.lib;bundle-version="2.37.0",
  org.eclipse.jdt.core;bundle-version="3.29.0",
  org.eclipse.emf.common;bundle-version="2.24.0",
- org.eclipse.emf.ecore;bundle-version="2.26.0"
+ org.eclipse.emf.ecore;bundle-version="2.26.0",
+ org.eclipse.xtext.jdt.facade;bundle-version="2.37.0"
 Export-Package: org.eclipse.xtext.java;version="2.37.0",
  org.eclipse.xtext.java.resource;version="2.37.0"
 Import-Package: org.apache.log4j;version="1.2.24"
