+++ org.eclipse.xtext.jdt.facade/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+encoding/<project>=ISO-8859-1
+encoding/src=ISO-8859-1
