+++ org.eclipse.xtext.jdt.facade/META-INF/MANIFEST.MF
@@ -1,0 +1,12 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Xtext JDT Facade
+Bundle-SymbolicName: org.eclipse.xtext.jdt.facade
+Bundle-Version: 2.37.0.qualifier
+Bundle-Vendor: Eclipse Xtext
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Require-Bundle: org.eclipse.jdt.core;bundle-version="3.29.0"
+Export-Package: org.eclipse.xtext.jdt.facade;version="2.37.0";x-friends:="org.eclipse.xtext.java"
+Import-Package: org.apache.log4j;version="1.2.24"
+Automatic-Module-Name: org.eclipse.xtext.jdt.facade
+Eclipse-SourceReferences: eclipseSourceReferences
