+++ org.eclipse.xtext.jdt.facade/build.properties
@@ -1,0 +1,5 @@
+source.. = src
+bin.includes = META-INF/,\
+               .,\
+               about.html
+src.includes = about.html
