+++ org.eclipse.xtext.runtime.feature/feature.xml
@@ -171,4 +171,18 @@ SPDX-License-Identifier: EPL-2.0
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="org.eclipse.xtext.jdt.facade"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
+   <plugin
+         id="org.eclipse.xtext.jdt.facade.source"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
